Initialization of Customer Database
