+++ userservice/src/main/java/com/user/userservice/CustomerDao.java
@@ -3,7 +3,7 @@ package com.user.userservice;
 import org.springframework.stereotype.Component;
 
 @Component
-public class UserDao {
+public class CustomerDao {
     //This file will interact with the database and provide functions such as:
