+++ userservice/src/main/java/com/user/userservice/CustomerRest.java
@@ -3,7 +3,7 @@ package com.user.userservice;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class UserRest {
+public class CustomerRest {
     
     //Provides the REST api for the user service
 }
