+++ userservice/src/main/resources/application.properties
@@ -1,1 +1,8 @@
-
+spring.datasource.url=jdbc:h2:file:./data/user
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.defer-datasource-initialization=true
+spring.h2.console.enabled=true
+spring.sql.init.mode=always
