+++ userservice/src/main/resources/data.sql
@@ -1,0 +1,5 @@
+ INSERT INTO customer (firstName, lastName, gender, email, phone) VALUES ('Brett', 'Renz', 'M', 'b@b.com', 123);
+ INSERT INTO customer (firstName, lastName, gender, email, phone) VALUES ('Joseph', 'Kerstetter', 'M', 'b@b.com', 123);
+ INSERT INTO customer (firstName, lastName, gender, email, phone) VALUES ('Taylor', 'Harbin', 'M', 'b@b.com', 123);
+ INSERT INTO customer (firstName, lastName, gender, email, phone) VALUES ('Ikechukwu', 'Amechi', 'M', 'b@b.com', 123);
+ INSERT INTO customer (firstName, lastName, gender, email, phone) VALUES ('Milcyas', 'Amualto', 'M', 'b@b.com', 123);
