+++ userservice/src/main/resources/schema.sql
@@ -1,0 +1,8 @@
+CREATE TABLE IF NOT EXISTS Customer (
+    id INTEGER PRIMARY KEY AUTO_INCREMENT,
+    firstName VARCHAR(50) NOT NULL,
+    lastName VARCHAR(50) NOT NULL,
+    gender VARCHAR(10) NOT NULL,
+    email VARCHAR(100) NOT NULL,
+    phone INTEGER NOT NULL
+);
